+++ prep_dist.sh
@@ -176,11 +176,12 @@ if asksure; then
     echo "Building for $plat"
     echo
     sudo rm -fR dist-$plat/*
-    if [ "$plat" = 'Darwin' ]; then
-        cxfreeze3 -s -c -OO etm --icon=etmTk/etmlogo.icns --target-dir dist-$plat/etmtk-${tag}-freeze-$plat
-    else
-        cxfreeze -OO etm --target-dir dist-$plat/etmtk-${tag}-freeze-$plat
-    fi
+    python3 setup_cxfreeze.py build
+#    if [ "$plat" = 'Darwin' ]; then
+#        cxfreeze3 -s -c -OO etm --icon=etmTk/etmlogo.icns --target-dir dist-$plat/etmtk-${tag}-freeze-$plat
+#    else
+#        cxfreeze -OO etm --target-dir dist-$plat/etmtk-${tag}-freeze-$plat
+#    fi
     cd dist-$plat
     tar czf etmtk-${tag}-freeze-$plat.tar.gz etmtk-${tag}-freeze-$plat
 #    zip -r etmtk-${tag}-freeze-UBUNTU.zip etmtk-${tag}-freeze-UBUNTU